Moved DES assign statements function and fixed error when no DES block
